requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.19.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
jinja2>=3.1.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
markdown>=3.5.0,<4.0.0
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
openai>=1.0.0,<2.0.0
